requirements.txt
pillow
streamlit
pandas
albumentations
pytorch 
torchvision
numpy